Fix ONR-SS seg fault issue
inDataWidth of post proc kernel should be equal to
width of output tidl tensor. Previously it was equal to
width of input tidl tensor.
In model ONR-SS-8610-deeplabv3lite-mobv2-ade20k32-512x512
the input and output tensor dimensions are not same,
as compared to other segmentation models where input and
output tensor dimensions are same, thereby exposing this
bug in application

Signed-off-by: Shubham Jain <a0492788@ti.com>